requirements.txt
alembic==1.4.1
appdirs==1.4.3
attrs==19.3.0
audioread==2.1.8
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
configparser==4.0.2
cycler==0.10.0
databricks-cli==0.9.1
decorator==4.4.2
docker==4.2.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
gitdb==4.0.2
GitPython==3.1.0
gorilla==0.3.0
gql==0.2.0
graphql-core==1.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mlflow==1.7.1
more-itertools==8.2.0
numba==0.48.0
numpy==1.18.2
nvidia-ml-py3==7.352.0
orion==0.1.7
packaging==20.3
pandas==1.0.3
pathtools==0.1.2
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
promise==2.3
protobuf==3.11.3
psutil==5.7.0
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pymongo==3.10.1
pyparsing==2.4.6
PyPDF2==1.26.0
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.23.0
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sentry-sdk==0.14.3
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
smmap==3.0.1
SoundFile==0.10.3.post1
SQLAlchemy==1.3.13
sqlparse==0.3.1
subprocess32==3.5.4
tabulate==0.8.6
torch==1.4.0
torchaudio==0.4.0
torchsampler==0.1
torchvision==0.5.0
tqdm==4.43.0
urllib3==1.25.8
wandb==0.8.30
watchdog==0.10.2
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
xlrd==1.2.0
zipp==3.1.0
